% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWORD.R
\name{predict.sword_flat}
\alias{predict.sword_flat}
\title{Predict from a fitted SWORD forest}
\usage{
\method{predict}{sword_flat}(object, newdata, ...)
}
\arguments{
\item{object}{\code{sword_flat} object from \code{\link{SWORD}}.}

\item{newdata}{data.frame of predictors.}

\item{...}{ignored.}
}
\value{
Numeric vector of predictions.
}
\description{
Predict from a fitted SWORD forest
}
\examples{
\donttest{
set.seed(1)
X <- data.frame(matrix(rnorm(150 * 5), 150, 5,
                dimnames = list(NULL, paste0("x", 1:5))))
y <- X$x1 * 2 - X$x2 + rnorm(150)
forest <- SWORD(X, y, m = 10, OOB = FALSE, verbose = FALSE)
preds  <- predict(forest, X[1:5, ])
}
}
