% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWORD.R
\name{predict.tors_flat}
\alias{predict.tors_flat}
\title{Predict from a fitted TORS tree}
\usage{
\method{predict}{tors_flat}(object, newdata, ...)
}
\arguments{
\item{object}{\code{tors_flat} object from \code{\link{TORS}}.}

\item{newdata}{data.frame of predictors.}

\item{...}{ignored.}
}
\value{
Numeric vector of predictions.
}
\description{
Predict from a fitted TORS tree
}
\examples{
set.seed(1)
X <- data.frame(matrix(rnorm(150 * 5), 150, 5,
                dimnames = list(NULL, paste0("x", 1:5))))
y <- X$x1 * 2 - X$x2 + rnorm(150)
tree  <- TORS(X, y, nmin = 10, cp = 0.01)
preds <- predict(tree, X[1:5, ])
}
